refactor: 전역에서 쓸 타입 분리
